Added Adapter.cpp
Motivation: We need to check if Peg fits into hole. hole is of round type. peg is of multiple types such as round and square. Roundpeg is compatible with round hole but squarepeg is not compatible with roundhole which requires roundpeg type. 
Added adapter design pattern. Commented line in main function illustrates incompatibility between squarepeg and roundpeg classes.
For which we are using adapter which creates squarepeg object and inherits roundpeg class which in turn enables us to convert squarepeg object into roundpeg via roundpeg interface